Allow users access to their root folders even if not written by them.
We're using a user's root's oid as their user id.  This makes it easy to:

a) pick user ids :)

b) and let them access their root after it was created by the root user.

Also, use the fact that the root certs objet is now a Cert object,
which let's us only look at a single object in the acceptor.