Refactored to avoid z64
I'd like to figure out some way in the not too distant future to make
the root object more easily virtualizable in ZODB so that ZeroDB
doesn't have to use custom DB and Connection classes.  I;m wondering
of there's some way to redirect z64 to the user's root folder.  In any
case, I think it's best not to rely n accessing data at z64.

For this reason, I decided to move the admin data into an Admin object
at oid 1. (This is ensured in init_db.) Then, for accessing admin
data, we go straight to oid 1, without traversing 0.

Also:

- It's nicer having the admin functions living in an Admin object.

- Users ids and their root folders' oids are now always the same,
  including for the admin user.  This provides some conceptual and
  code simplification. (At the small cost of arranging for and relying
  on the root users oid being 2. :) )

- Changed users_by_der to uids {der -> uid), since the custom
  ZEOStorage only needs the user id.

- Added a user name -> user mapping to support some upcoming admin functions.

- Simplified the way the acceptor opens it's DB. It's cute that having
  the user id as a suffic doesn't break unnpickling.  We may not want
  to rely on this in the long run.

  (I continue to mull alternate serialization formats.)